fix(data-model-schema): the canonical selectors are canonical, and the package is acyclic
Two things, the second of which is what makes the first safe to state.

## The canonical selectors

`REJECTING_SELECTOR` and `DEFAULT_SELECTOR` are documented as canonical but
were hand-frozen rather than registered in the cache `internPathSelector()`
keys on, so which instance ended up canonical was decided by call order.
Interning a constant first made it canonical and later equivalents collapsed
onto it; interning any equivalent first made *that* one canonical and left the
exported constant a duplicate for the life of the process. Both orders
measured. The identity-keyed caches `internPathSelector()` exists to serve --
`hashOf()`'s frozen-object `WeakMap`, `MapSetStringToPathSelectors`, the cycle
trackers -- could therefore hold two wrappers for one selector.

Both constants are now defined *through* `internPathSelector()`, so each claims
its slot at module load, before any caller can build an equivalent. The test
builds its equivalents from literals rather than handing a constant back to
`internPathSelector()`, which would let the constant claim the slot on the spot
and pass whether or not the module seeded anything.

## The module cycle

`schema-intern.ts` and `schema-utils.ts` imported each other. The single edge
from the first to the second was `toDeepFrozenSchema()`, and the only thing
tying that function to interning was an `isInternedSchema()` early-out --
redundant, because both arms below it already recognize an interned schema.
`deepFreeze()` returns a `boolean` or `undefined` untouched and identity-returns
an already-deep-frozen object through its cache, and `cloneIfNecessary()`
identity-passes an already-deep-frozen value. So the early-out is gone and the
function reads as the two cases it always was.

That freed `toDeepFrozenSchema()` to leave `schema-utils.ts`, which is what
actually cuts the edge. It goes to `schema-copy.ts` along with
`cloneSchemaMutable()`, the package's other copying operation, which was
already intern-free. `basic-schemas.ts` takes the lazily-interned
`{ type: <name> }` table, `getBasicSchema()`, `schemaForValueType()` and
`emptySchemaObject()`. What is left in `schema-utils.ts` is the schema
*rewriting* -- `schemaWithProperties()`, `schemaWithoutProperties()` -- plus
`isNontrivialSchema()` and `internSchemaPairAsKey()`.

The package's modules now form a DAG, `schema-copy.ts` a leaf and
`schema-intern.ts` above it. That is the invariant the canonical selectors rest
on: `path-selector.ts` can intern at load only because `schema-intern.ts` is
guaranteed to have finished first, and the module comment says so, along with
what would forfeit it.

Test bodies move with their subjects and are otherwise unchanged.
