+++ packages/data-model-schema/src/schema-utils.ts
@@ -1,16 +1,8 @@
 /** Runtime utilities for working with JSONSchema values. */
 
-import type {
-  JSONSchema,
-  JSONSchemaObj,
-  JSONSchemaTypes,
-  MutableJSONSchemaObj,
-} from "@commonfabric/api";
+import type { JSONSchema, JSONSchemaObj } from "@commonfabric/api";
 
-import { deepFreeze } from "@commonfabric/data-model/deep-freeze";
 import {
-  cloneIfNecessary,
-  FabricPrimitive,
   type FabricValue,
   shallowMutableClone,
 } from "@commonfabric/data-model/fabric-value";
@@ -19,30 +11,7 @@ import {
   internSchemaAsTaggedHashString,
   isInternedSchema,
 } from "./schema-intern.ts";
-import { schemaTypeOfFabricPrimitive } from "./schemaTypeOfFabricPrimitive.ts";
-
-/**
- * Map from `JSONSchema` type names (and special names) to corresponding
- * interned schemas. Populated lazily.
- */
-const BASIC_SCHEMAS: Record<string, JSONSchemaObj> = {};
-
-/**
- * Helper for `schemaForValueType()` and `emptySchemaObject()`, which does
- * the lookup and interning as necessary.
- */
-function getBasicSchema(key: string) {
-  const found = BASIC_SCHEMAS[key];
-
-  if (found) {
-    return found;
-  } else {
-    const result = BASIC_SCHEMAS[key] = internSchema({
-      type: key as JSONSchemaTypes,
-    });
-    return result;
-  }
-}
+import { toDeepFrozenSchema } from "./schema-copy.ts";
 
 /**
  * Indicates if the given (nullable) schema is in fact a non-trivial schema. A
@@ -65,75 +34,6 @@ export function isNontrivialSchema(
   }
 
   return Object.keys(schema).length !== 0;
-}
-
-/**
- * Returns a deep-frozen copy of (or reference to) a JSONSchema, returning
- * primitives as-is.
- *
- * - When `canShare` is `true`, the input `schema` is allowed to be modified,
- *   including freezing it in place and returning it directly. Use this when the
- *   caller owns the object referred to by `schema` and no other code will
- *   attempt to mutate it.
- *
- * - When `canShare` is `false`, the `schema` is cloned first if not already
- *   deep-frozen, so that the original is not modified.
- *
- * As with other schema functions, this one accepts `undefined` for use when
- * it's possible for a schema value to be missing or optional.
- *
- * Note: Use `internSchema()` in preference to this function, which can cost a
- * little more to run but which will save both time and memory when the schema
- * in question is reused.
- */
-export function toDeepFrozenSchema<T extends JSONSchema | undefined>(
-  schema: T,
-  canShare: boolean = false,
-): T {
-  // No need to do any work given an interned schema (including `boolean`s.)
-  if (isInternedSchema(schema)) {
-    return schema;
-  }
-
-  // After the boolean check, `schema` is necessarily a `JSONSchemaObj`. We use
-  // a local `schemaObj` variable so TypeScript can track the object-only type,
-  // then cast back to `T` on return.
-  const schemaObj = schema as Exclude<T, boolean>;
-
-  if (canShare) {
-    // The caller indicated that we get to freeze the result, so just do that.
-    // The call to `deepFreeze()` is a relatively inexpensive no-op if
-    // `schemaObj` is in fact already deep-frozen.
-    return deepFreeze(schemaObj);
-  } else {
-    // The caller indicated that the original `schema` has to be left alone, so
-    // make a deep-frozen clone of it. As with `deepFreeze()`, if it turns out
-    // `schemaObj` is already deep-frozen, the call is a relatively inexpensive
-    // no-op.
-    return cloneIfNecessary(schemaObj) as T;
-  }
-}
-
-/**
- * Returns a mutable object copy of a JSONSchema. Boolean schemas (`true` and
- * `false`) and `undefined` are converted to their object-form equivalents:
- * `undefined` and `true` become `{}` (accept any value), `false` becomes
- * `{ not: true }` (reject all values).
- *
- * @param deep When `true`, nested objects are recursively cloned (deep copy).
- *   Defaults to `false` (shallow copy). Pass `true` when the caller intends to
- *   mutate nested properties.
- */
-export function cloneSchemaMutable(
-  schema: JSONSchema | undefined,
-  deep: boolean = false,
-): MutableJSONSchemaObj {
-  if (schema === undefined) return {};
-  if (typeof schema === "boolean") return schema ? {} : { not: true };
-  return cloneIfNecessary(schema, {
-    frozen: false,
-    deep,
-  }) as MutableJSONSchemaObj;
 }
 
 /**
@@ -235,70 +135,6 @@ export function schemaWithoutProperties(
     // Note: We still have to deep-freeze in the `!copy` case, though it will be
     // a no-op if `schema` was already deep-frozen (including interned).
     return toDeepFrozenSchema(schema);
-  }
-}
-
-/**
- * Gets the basic `{ type: name }` schema for a given value. Returns `undefined`
- * if there is no well-defined type for the value. The result is always interned
- * (and frozen).
- *
- * **Note:** `undefined` (as a value) is in an "intermediate" state in the
- * codebase as of this writing, and _this_ function treats it as not having a
- * well-defined type.
- */
-export function schemaForValueType(
-  value: FabricValue,
-): JSONSchemaObj | undefined {
-  // TODO(danfuzz): This is a place that will need to get smarter once we
-  // actually want to accept values beyond what's strictly allowed in JSON. This
-  // notably includes `undefined` and all the other non-plain-object
-  // `FabricValue` possibilities.
-
-  const type = typeof value;
-  switch (type) {
-    case "object": {
-      if (value === null) {
-        return getBasicSchema("null");
-      } else if (Array.isArray(value)) {
-        return getBasicSchema("array");
-      } else if (value instanceof FabricPrimitive) {
-        // A `FabricPrimitive` gets its specific type name (e.g.
-        // "FabricBytes") rather than "object": it is an opaque leaf, so
-        // "object" would invite structural keywords that cannot apply.
-        return getBasicSchema(schemaTypeOfFabricPrimitive(value));
-      }
-      break;
-    }
-
-    case "number": {
-      if (Number.isInteger(value)) {
-        return getBasicSchema("integer");
-      }
-      break;
-    }
-
-    case "bigint":
-    case "symbol":
-    case "undefined": {
-      // Not accepted yet, even though the intention is to accept most or all
-      // of these.
-      return undefined;
-    }
-  }
-
-  return getBasicSchema(type);
-}
-
-/** Gets the standard interned empty schema _object_, a literal `{}`. */
-export function emptySchemaObject() {
-  const key = "emptySchema";
-  const found = BASIC_SCHEMAS[key];
-  if (found) {
-    return found;
-  } else {
-    const result = BASIC_SCHEMAS[key] = internSchema({});
-    return result;
   }
 }
