+++ packages/data-model-schema/test/SchemaAndHash.test.ts
@@ -16,7 +16,7 @@ import { expect } from "@std/expect";
 
 import { FabricHash } from "@commonfabric/data-model/fabric-primitives";
 import { SchemaAndHash } from "@/SchemaAndHash.ts";
-import { toDeepFrozenSchema } from "@/schema-utils.ts";
+import { toDeepFrozenSchema } from "@/schema-copy.ts";
 
 // Two distinct non-empty byte arrays so tests can observe variance.
 const HASH_A = new FabricHash(
